fix(practice): Header 컴포넌트 분리
Header 컴포넌트를 App.jsx 파일에서 분리하였습니다.
