+++ practice/01-starting-project/src/App.jsx
@@ -1,27 +1,6 @@
-import reactImg from './assets/react-core-concepts.png';
 import componentsImg from './assets/Components.png';
 import { CORE_CONCEPTS } from './data.js';
-
-const reactDescriptions = ['Fundamental', 'Crucial', 'Core'];
-
-function genRandomInt(max) {
-  return Math.floor(Math.random() * (max + 1));
-}
-
-function Header() {
-
-  const description = reactDescriptions[genRandomInt(reactDescriptions.length - 1)];
-
-  return (
-    <header>
-      <img src={reactImg} alt="Stylized atom" />
-      <h1>React Essentials</h1>
-      <p>
-        {description} React concepts you will need for almost any app you are going to build!
-      </p>
-    </header>
-  );
-}
+import Header from'./components/Header.jsx';
 
 function CoreConcept({image, title, description}) {
   return <li>
